# internal/server/swagger/swagger.yaml
basePath: /
definitions:
...
  files.EditorType:
    enum:
    - filter-editor
    - list-editor
    - todo-editor
    - index-editor
    - codemirror-editor
    type: string
    x-enum-varnames:
    - EditorTypeFilter
    - EditorTypeList
    - EditorTypeTodo
    - EditorTypeIndex
    - EditorTypeCodeMirror
...
  files.File:
    properties:
      metadata:
        $ref: '#/definitions/files.Metadata'
      name:
        type: string
      path:
        type: string
    type: object
...
  files.Metadata:
    properties:
      ancestor:
        description: auto
        items:
          type: string
        type: array
      collection:
        description: auto
        type: string
      conflictFile:
        description: auto
        type: string
      conflictOf:
        description: auto
        type: string
      createdAt:
        description: auto
        type: string
      editor:
        allOf:
        - $ref: '#/definitions/files.EditorType'
        description: manual
      folders:
        description: auto
        items:
          type: string
        type: array
      kanbanAddedAt:
        description: auto
        type: string
      kanbanMovedAt:
        description: auto
        type: string
      kids:
        description: auto
        items:
          type: string
        type: array
      lastEdited:
        description: auto
        type: string
      linksToHere:
        description: auto
        items:
          type: string
        type: array
      parents:
        description: manual
        items:
          type: string
        type: array
      path:
        description: auto
        type: string
      references:
        description: manual
        items:
          $ref: '#/definitions/files.Reference'
        type: array
      related:
        description: auto
        items:
          type: string
        type: array
      size:
        description: auto
        type: integer
      tags:
        description: manual
        items:
          type: string
        type: array
      title:
        description: auto
        type: string
      usedLinks:
        description: auto
        items:
          type: string
        type: array
    type: object
  files.Reference:
    properties:
      addedAt:
        type: string
      description:
        description: why this link was added
        type: string
      url:
        type: string
    type: object
...
  filter.Result:
    properties:
      files:
        items:
          $ref: '#/definitions/files.File'
        type: array
      filter_count:
        type: integer
      logic:
        type: string
      total:
        type: integer
    type: object
...
host: localhost:1324
info:
  contact: {}
  description: KNOV API \n http://localhost:1324
  title: Knov API
  version: "1.0"
paths:
...
  /api/filters:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Filter files based on metadata criteria with configurable logic
        and display
      parameters:
      - description: Metadata field names
        in: formData
        name: metadata[]
        type: array
      - description: Filter operators (equals, contains, greater, less, in)
        in: formData
        name: operator[]
        type: array
      - description: Filter values
        in: formData
        name: value[]
        type: array
      - description: Filter actions (include, exclude)
        in: formData
        name: action[]
        type: array
      - default: and
        description: Logic operator (and/or)
        in: formData
        name: logic
        type: string
      - default: list
        description: Display type (list, cards, dropdown, table)
        in: formData
        name: display
        type: string
      - default: 50
        description: Maximum number of results
        in: formData
        name: limit
        type: integer
      produces:
      - application/json
      - text/html
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/filter.Result'
      summary: Filter files by metadata
      tags:
      - filter
  /api/filters/{id}:
    delete:
      description: Delete a filter from config storage and its metadata
      parameters:
      - description: filter id
        in: path
        name: id
        required: true
        type: string
      produces:
      - text/html
      responses:
        "200":
          description: deleted
          schema:
            type: string
      summary: Delete filter
      tags:
      - filter
...
  /api/filters/criteria-row:
    get:
      consumes:
      - application/x-www-form-urlencoded
      description: Get HTML for a new filter criteria row
      parameters:
      - description: Row index
        in: formData
        name: row_index
        type: integer
      produces:
      - text/html
      responses:
        "200":
          description: filter criteria row html
          schema:
            type: string
      summary: Get filter criteria row
      tags:
      - filter
  /api/filters/save:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Save filter configuration to config storage
      parameters:
      - description: Filter identifier (name)
        in: formData
        name: filterid
        required: true
        type: string
      - description: Metadata field names
        in: formData
        name: metadata[]
        type: array
      - description: Filter operators (equals, contains, greater, less, in)
        in: formData
        name: operator[]
        type: array
      - description: Filter values
        in: formData
        name: value[]
        type: array
      - description: Filter actions (include, exclude)
        in: formData
        name: action[]
        type: array
      - default: and
        description: Logic operator (and/or)
        in: formData
        name: logic
        type: string
      produces:
      - text/html
      responses:
        "200":
          description: success message
          schema:
            type: string
      summary: Save filter configuration
      tags:
      - filter
  /api/filters/value-input:
    get:
      consumes:
      - application/x-www-form-urlencoded
      description: Get HTML for filter value input based on metadata field type
      parameters:
      - description: Metadata field name
        in: formData
        name: metadata
        required: true
        type: string
      - description: Row index
        in: formData
        name: row_index
        required: true
        type: integer
      - description: Current value
        in: formData
        name: value
        type: string
      produces:
      - text/html
      responses:
        "200":
          description: filter value input html
          schema:
            type: string
      summary: Get filter value input
      tags:
      - filter
...
  /api/kanban/{board}/filter:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Filters the kanban board using the full filter form, scoped to
        the board's folder
      parameters:
      - description: Board slug
        in: path
        name: board
        required: true
        type: string
      produces:
      - application/json
      - text/html
      responses:
        "200":
          description: kanban board html
          schema:
            type: string
      summary: Apply advanced filter to kanban board
      tags:
      - kanban
...
swagger: "2.0"